Move image loading to adapter
