+++ src/packager/web/adapter.js
@@ -1,4 +1,6 @@
 import assetCache from './cache';
+import request from '../../common/request';
+import {readAsURL, readAsArrayBuffer} from '../../common/readers';
 
 class WebAdapter {
   getCachedAsset (asset) {
@@ -7,6 +9,50 @@ class WebAdapter {
 
   async cacheAsset (asset, result) {
     await assetCache.set(asset, result);
+  }
+
+  getAppIcon (file) {
+    if (!file) {
+      return request({
+        url: defaultIcon,
+        type: 'arraybuffer'
+      });
+    }
+    // Convert to PNG
+    if (file.type === 'image/png') {
+      return readAsArrayBuffer(file);
+    }
+    return new Promise((resolve, reject) => {
+      const url = URL.createObjectURL(file);
+      const image = new Image();
+      image.onload = () => {
+        image.onload = null;
+        image.onerror = null;
+        const canvas = document.createElement('canvas');
+        const ctx = canvas.getContext('2d');
+        if (!ctx) {
+          reject(new Error('Cannot get rendering context for icon conversion'));
+          return;
+        }
+        canvas.width = image.width;
+        canvas.height = image.height;
+        ctx.drawImage(image, 0, 0);
+        canvas.toBlob((blob) => {
+          URL.revokeObjectURL(url);
+          resolve(readAsArrayBuffer(blob));
+        });
+      };
+      image.onerror = () => {
+        image.onload = null;
+        image.onerror = null;
+        reject(new Error('Cannot load icon'));
+      };
+      image.src = url;
+    });
+  }
+
+  readAsURL (file) {
+    return readAsURL(file);
   }
 }
